feat(cli): move all subcommand handlers into its own module
